Implement CRUD controllers

+++ src/main/java/repo/EmployeeRepo.java
@@ -12,5 +12,5 @@ import model.Employee;
 // }
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepo extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepo extends JpaRepository<Employee, Long> {
 }
